Move basic auth routes to a separate file
